+++ src/main/java/com/example/shortform/service/PostService.java
@@ -1,5 +1,6 @@
 package com.example.shortform.service;
 
+import com.example.shortform.config.auth.PrincipalDetails;
 import com.example.shortform.domain.Challenge;
 import com.example.shortform.domain.Comment;
 import com.example.shortform.domain.ImageFile;
@@ -41,13 +42,14 @@ public class PostService {
 
     @Transactional
     public ResponseEntity<?> writePost(Long challengeId,
-                                       @RequestPart("post") PostRequestDto requestDto,
-                                       @RequestPart(value = "image",required = false) MultipartFile multipartFile) throws IOException {
+                                       PostRequestDto requestDto,
+                                       MultipartFile multipartFile,
+                                       PrincipalDetails principalDetails) throws IOException {
         Challenge challenge = challengeRepository.findById(challengeId).orElseThrow(
                 () -> new NullPointerException("챌린지가 존재하지 않습니다.")
         );
 
-        Post post = postRepository.save(requestDto.toEntity(challenge));
+        Post post = postRepository.save(requestDto.toEntity(challenge, principalDetails.getUser()));
 
         ImageFile imageFile = imageFileService.upload(multipartFile, post);
 
@@ -57,19 +59,32 @@ public class PostService {
     }
 
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, PrincipalDetails principalDetails) {
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new NullPointerException("게시글이 존재하지 않습니다.")
         );
+
+        if (!principalDetails.getUser().getId().equals(post.getUser().getId())) {
+            throw new IllegalArgumentException("작성자만 삭제할 수 있습니다.");
+        }
 
         postRepository.deleteById(postId);
     }
 
     @Transactional
-    public ResponseEntity<?> modifyPost(Long postId, PostRequestDto requestDto) {
+    public ResponseEntity<?> modifyPost(Long postId,
+                                        PostRequestDto requestDto,
+                                        PrincipalDetails principalDetails,
+                                        MultipartFile multipartFile) throws IOException {
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new NullPointerException("인증 게시글이 존재하지 않습니다.")
         );
+
+        ImageFile imageFile = imageFileService.upload(multipartFile, post);
+
+        if (!principalDetails.getUser().getId().equals(post.getUser().getId())) {
+            throw new IllegalArgumentException("작성자만 수정할 수 있습니다.");
+        }
 
         post.update(requestDto);
 
@@ -82,30 +97,33 @@ public class PostService {
                 () -> new NullPointerException("챌린지가 존재하지 않습니다.")
         );
 
-        List<Post> postList = postRepository.findAllByChallenge(challenge);
+        List<Post> postList = postRepository.findAllByChallengeId(challengeId);
         List<PostResponseDto> responseDtoList = new ArrayList<>();
-        List<CommentResponseDto> commentDetailList = new ArrayList<>();
+
 
         for (Post post : postList) {
-            List<Comment> commentList = commentRepository.findAllByPost(post);
+            List<CommentResponseDto> commentDetailList = new ArrayList<>();
+            List<Comment> commentList = commentRepository.findAllByPostId(post.getId());
             for (Comment comment : commentList) {
                 CommentResponseDto commentDetailResponseDto = comment.toResponse();
-                String commentCreatedAt = commentDetailResponseDto.getCreatedAt();
+                String commentCreatedAt = comment.getCreatedAt().toString();
                 String year = commentCreatedAt.substring(0,4) + "년";
                 String month = commentCreatedAt.substring(5,7) + "월";
                 String day = commentCreatedAt.substring(8,10) + "일";
                 String time = commentCreatedAt.substring(11,19);
                 commentCreatedAt = year + month + day + time;
-                commentDetailList.add(commentDetailResponseDto.setCreatedAt(commentCreatedAt));
+                commentDetailResponseDto.setCreatedAt(commentCreatedAt);
+                commentDetailList.add(commentDetailResponseDto);
             }
             PostResponseDto postResponseDto = post.toResponse(commentDetailList);
-            String postCreatedAt = postResponseDto.getCreatedAt();
+            String postCreatedAt = post.getCreatedAt().toString();
             String year = postCreatedAt.substring(0,4) + "년";
             String month = postCreatedAt.substring(5,7) + "월";
             String day = postCreatedAt.substring(8,10) + "일";
             String time = postCreatedAt.substring(11,19);
             postCreatedAt = year + month + day + time;
-            responseDtoList.add(postResponseDto.setCreatedAt(postCreatedAt));
+            postResponseDto.setCreatedAt(postCreatedAt);
+            responseDtoList.add(postResponseDto);
         }
 
         return ResponseEntity.ok(responseDtoList);
